`data_views`: Move `@kbn/config-schema` usage from `common` to `server`
