+++ src/plugins/data_views/server/rest_api_routes/public/update_data_view.ts
@@ -13,11 +13,7 @@ import { DataViewSpecRestResponse } from '../route_types';
 import { DataViewsService } from '../../../common/data_views';
 import { DataViewSpec } from '../../../common/types';
 import { handleErrors } from './util/handle_errors';
-import {
-  fieldSpecSchema,
-  runtimeFieldSchema,
-  serializedFieldFormatSchema,
-} from '../../../common/schemas';
+import { fieldSpecSchema, runtimeFieldSchema, serializedFieldFormatSchema } from '../../schemas';
 import { dataViewSpecSchema } from '../schema';
 import type {
   DataViewsServerPluginStartDependencies,
